Two version

+++ src/main/java/ru/netology/stats/StatsService.java
@@ -3,8 +3,8 @@ package ru.netology.stats;
 public class StatsService {
 
 
-    public double minSales(double[] sales) {
-        double minMonth = 0; // номер месяца с минимальными продажами среди просмотренных ранее
+    public int minSales(int[] sales) {
+        int minMonth = 0; // номер месяца с минимальными продажами среди просмотренных ранее
 
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] <= sales[(int) minMonth]) { // значит, в рассматриваемом i-м месяце продаж меньше
@@ -15,7 +15,7 @@ public class StatsService {
         return minMonth + 1; // месяца нумеруются с 1, а индексы массива с 0, нужно сдвинуть ответ на 1
     }
 
-    public double maxSales(double[] sales) {
+    public int maxSales(int[] sales) {
         int maxMonth = 0;
 
         for (int i = 0; i < sales.length; i++) {
@@ -26,23 +26,23 @@ public class StatsService {
         return maxMonth + 1;
 
     }
-    public double salesAmount(double[] sales) {
-        double amount = 0;
+    public int salesAmount(int[] sales) {
+        int amount = 0;
         for (int i = 0; i < sales.length; i++) {
             amount = amount + sales[i];
         }
         return amount;
     }
 
-    public double averageAmount(double[] sales) {
-        double amount = salesAmount(sales);
-        double salesLength = sales.length;
-        double averageSalesAmount = amount / salesLength;
+    public int averageAmount(int[] sales) {
+        int amount = salesAmount(sales);
+        int salesLength = sales.length;
+        int averageSalesAmount = amount / salesLength;
         return averageSalesAmount;
     }
 
-    public int SalesAboveAverage (double[] sales) {
-        double medium = averageAmount(sales);
+    public int SalesAboveAverage (int[] sales) {
+        int medium = averageAmount(sales);
         int count = 0;
         for (int i = 0; i <sales.length; i++) {
             if (sales[i] > medium) {
@@ -52,8 +52,8 @@ public class StatsService {
         return count;
     }
 
-    public int SalesOfLessThanAverage (double[] sales) {
-        double medium = averageAmount(sales);
+    public int SalesOfLessThanAverage (int[] sales) {
+        int medium = averageAmount(sales);
         int count = 0;
         for (int i = 0; i <sales.length; i++) {
             if (sales[i] < medium) {
